backend/python_recommender/requirements.txt
flask==3.0.3
langchain-google-genai==2.1.5
python-dotenv==1.0.1
